Reset Filters button enables if keyword, tags, or both are used otherwise disabled
