+++ civictechprojects/static/css/ProjectTagContainer.css
@@ -12,6 +12,7 @@
 
 }
 
+/*if the reset search button is used elsewhere, split this out of ProjectTagContainer.css */
 button.ResetSearch {
   background-color: #86c232;
   border: 0;
@@ -22,4 +23,7 @@ button.ResetSearch {
   font-size: 14px;
   float: right;
   padding: 10px 30px;
+}
+button.ResetSearch:disabled {
+  background-color: #dddddd;
 }
